Custom Exception and Middleware are added.

+++ UdemyPractice/src/BLL/Services/IDepartmentService.cs
@@ -1,10 +1,12 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.Threading.Tasks;
 using BLL.Request;
 using DLL.Model;
 using DLL.Repository;
 using DLL.UoW;
 using Microsoft.EntityFrameworkCore;
+using Utils.AppExceptions;
 
 namespace BLL.Services
 {
@@ -37,7 +39,14 @@ namespace BLL.Services
 
         public async Task<Department> GetAAsync(string code)
         {
-            return await _departmentRepository.FirstOrDefaultAsync(x => x.Code == code);
+            var department = await _departmentRepository.FirstOrDefaultAsync(x => x.Code == code);
+
+            if (department == null)
+            {
+                throw new ApplicationValidationException("The department was not found.");
+            }
+
+            return department;
         }
 
         public async Task<Department> InsertAsync(DepartmentInsertRequestModel request)
@@ -49,7 +58,11 @@ namespace BLL.Services
             };
 
             await _departmentRepository.CreateAsync(department);
-            await _unitOfWork.Commit();
+
+            if (await _unitOfWork.Commit())
+            {
+                throw new ApplicationValidationException("Something went wrong. Please try again later.");
+            }
 
             return department;
         }
@@ -57,21 +70,54 @@ namespace BLL.Services
         public async Task<Department> Update(string code, DepartmentUpdateRequestModel request)
         {
             var department = await _departmentRepository.FirstOrDefaultAsync(x => x.Code == code);
-            var updateDepartment = new Department()
+            if (department == null)
             {
-                Name = request.Name
-            };
-            await _departmentRepository.Update(updateDepartment);
-            await _unitOfWork.Commit();
-            return department;
+                throw new ApplicationValidationException("The department was not found.");
+            }
+            
+            if (!string.IsNullOrWhiteSpace(request.Code))
+            {
+                var existsAlreadyCode = await _departmentRepository.FirstOrDefaultAsync(x => x.Code == request.Code);
+                if (existsAlreadyCode != null)
+                {
+                    throw new ApplicationValidationException("This code already exists in our system.");
+                }
+
+                department.Code = request.Code;
+            }
+            
+            if (!string.IsNullOrWhiteSpace(request.Name))
+            {
+                var existsAlreadyCode = await _departmentRepository.FirstOrDefaultAsync(x => x.Name == request.Name);
+                if (existsAlreadyCode != null)
+                {
+                    throw new ApplicationValidationException("This code already exists in our system.");
+                }
+
+                department.Name = request.Name;
+            }
+
+            await _departmentRepository.Update(department);
+            if (await _unitOfWork.Commit())
+            {
+                return department;
+            }
+            throw new ApplicationValidationException("Something went wrong. Please try again later.");
         }
 
         public async Task<Department> Delete(string code)
         {
             var department = await _departmentRepository.FirstOrDefaultAsync(x => x.Code == code);
+            if (department == null)
+            {
+                throw new ApplicationValidationException("The department was not found.");
+            }
             await _departmentRepository.Delete(department);
-            await _unitOfWork.Commit();
-            return department;
+            if (await _unitOfWork.Commit())
+            {
+                return department;
+            }
+            throw new ApplicationValidationException("Something went wrong. Please try again later.");
         }
 
         public async Task<bool> IsCodeExists(string code)
